Add messages for LinguaLibre.org menu
Currently, these three menu items are written as
pages in the MediaWiki space. This is an inconvenient
way to localize them, because they require administrator
permissions. This change will put them on translatewiki
with the rest of the messages.

All the existing translations from LinguaLibre.org
were copied with credit.

After merging and deploying this, several things
will have to be done to complete the change:
1. Change "menu-*" entries on https://lingualibre.org/wiki/MediaWiki:Sidebar
   to mwe-recwiz-menu-*. (The prefixes were added for consistency
   with the rest of the extension, and for ensuring global message
   key uniqueness.)
2. Delete all the Menu-* pages in the MediaWiki namespace. See here:
   https://lingualibre.org/wiki/Special:AllPages?from=Menu-&namespace=8